Move title & tab helpers to separate file
