+++ app/helpers/licenses_helper.rb
@@ -2,45 +2,11 @@
 
 # helpers for show License view
 module LicensesHelper
-  def license_index_tabs
-    [new_license_tab]
-  end
-
-  def license_show_title(license)
-    [
-      license.display_name,
-      license_title_id(license)
-    ].safe_join(" ")
-  end
-
-  # "Editing: Creative Commons Non-commercial v3.0 (#nnn)"  textilized
-  def license_edit_title(license)
-    capture do
-      concat("#{:EDITING.l}: ")
-      concat(license_show_title(license))
-    end
-  end
-
-  def license_title_id(license)
-    tag.span(class: "smaller") do
-      tag.span("#(#{license.id || "?"}):")
-    end
-  end
-
   def license_updated_at(license)
     if license.updated_at
       license.updated_at.strftime("%Y-%m-%d %H:%M:%S")
     else
       "nil"
     end
-  end
-
-  ##########
-
-  private
-
-  def new_license_tab
-    [:create_license_title.t, new_license_path,
-     { class: tab_id(__method__.to_s) }]
   end
 end
